% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{bolt2002Model}
\alias{bolt2002Model}
\title{Fit graded response model to MP model based on procedure by Bolt (2002)}
\usage{
bolt2002Model(
  x,
  items,
  theta = seq(-4, 4, length.out = 500),
  w = dnorm(theta)/sum(dnorm(theta))
)
}
\arguments{
\item{x}{Fitted \code{mxModel}, e.g., from \code{\link{simAnneal}}.}

\item{items}{Vector of indices indicating which items to perform procedure for.}

\item{theta}{Grid across theta over which to perform estimation for Bolt (2002)'s approach.}

\item{w}{Weights corresponding to \code{theta} for performing estimation for Bolt (2002)'s approach.}
}
\value{
A list with the following elements
}
\description{
Fit graded response model to MP model based on procedure by Bolt (2002)
}
\details{
This function goes item by item for the model in \code{x} and uses the procedure by Bolt (2002) to fit
  a graded response model as best it can to each item's (semi-parametric) response functions.
}
\section{Slots}{

\describe{
\item{\code{conv}}{Convergence code for each item returned from \code{nlminb}.}

\item{\code{gr.items}}{A list of item specifications from the graded model.}

\item{\code{pars}}{A list of item parameters from the fitted graded model for each item.}

\item{\code{mp.items}}{A list of item specifications from the MP model.}

\item{\code{mp.pars}}{A list of item parameters from the MP model for each item.}
}}

