% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{computek}
\alias{computek}
\title{Compute k (used in specifying polynomial order) for an item
This seems to work at auto-detecting k when the item is an MP item
I suppose if this is used on an item that isn't an MP item, it'll yield some value,
but that value sure won't be k?}
\usage{
computek(x, j)
}
\arguments{
\item{x}{- fitted mxModel or matrix that contains a "1" in some row indicating the value of k for each item}

\item{j}{- item index}
}
\description{
Compute k (used in specifying polynomial order) for an item
This seems to work at auto-detecting k when the item is an MP item
I suppose if this is used on an item that isn't an MP item, it'll yield some value,
but that value sure won't be k?
}
