% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{computekrec}
\alias{computekrec}
\title{computes k for all items and returns k as a vector}
\usage{
computekrec(x, ni)
}
\arguments{
\item{x}{blah}

\item{ni}{blah}
}
\description{
computes k for all items and returns k as a vector
}
