% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{extractItem}
\alias{extractItem}
\title{Extracts useful information about a single item from a fitted MP model and returns all of these values in a list}
\usage{
extractItem(x, j, mp = TRUE)
}
\arguments{
\item{x}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{j}{Item index}

\item{mp}{Whether the item in question is an MP model (otherwise, don't compute k)}
}
\value{
A list with the item parameters (\code{pars}), number of categories (\code{ncat}),
  item model generated from \code{rpf} (\code{spec}), and possibly the integer (\code{k}) that controls polynomial order.
}
\description{
Extracts useful information about a single item from a fitted MP model and returns all of these values in a list
}
\examples{
\donttest{
# TODO examples here
}
}
