% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{extractItemMx}
\alias{extractItemMx}
\title{This gets all item parameters (pars), number of categories (ncat), and k for an item
and returns all of these values in a list}
\usage{
extractItemMx(x, j, mp = TRUE)
}
\arguments{
\item{x}{- fitted model}

\item{j}{- item index}

\item{mp}{- whether the item in question is an MP model (otherwise, don't compute k)}
}
\description{
This gets all item parameters (pars), number of categories (ncat), and k for an item
and returns all of these values in a list
}
