% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{getk}
\alias{getk}
\title{Compute k (used in specifying polynomial order) for a single item.}
\usage{
getk(x, j)
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}} or k matrix that contains a "1" in some row indicating the value of k for each item (e.g., as generated by \code{\link{newkmat}}).}

\item{j}{Item index.}
}
\description{
Compute k (used in specifying polynomial order) for a single item.
}
\details{
This seems to work at auto-detecting k when the item is an MP item
I suppose if this is used on an item that isn't an MP item, it'll yield some value,
but that value sure won't be k.
}
