% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{getkrec}
\alias{getkrec}
\title{Compute k for all items from a fitted MP model.}
\usage{
getkrec(x, ni)
}
\arguments{
\item{x}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{ni}{Number of items.}
}
\description{
Compute k for all items from a fitted MP model.
}
\details{
Compute k for all items and returns k as a vector.
}
