% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{IAD}
\alias{IAD}
\title{Integrated Absolute Difference}
\usage{
IAD(T1, T2, w)
}
\arguments{
\item{T1}{Vector of values corresponding to traceline or information function 1.}

\item{T2}{Vector of values corresponding to traceline or information function 2.}

\item{w}{Vector of weights; usually these should sum to 1.}
}
\description{
Integrated Absolute Difference
}
\details{
Computes integrated absolute difference (IAD) between \code{T1} and \code{T2}, with weights for integration
from \code{w}. Note, this function is essentially the same as AAD from Bolt (2002), wABC from Edelen, Stuck, and Chandra (2015), and uDTF from Chalmers, Counsell, and Flora (2016).
}
