% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{infodif}
\alias{infodif}
\title{Difference in information functions or TIF}
\usage{
infodif(
  mod1,
  mod2,
  j1,
  j2,
  method = c("RIMSD", "IAD", "ISD"),
  theta = seq(-5, 5, 0.1),
  w = dnorm(theta)/sum(dnorm(theta))
)
}
\arguments{
\item{mod1}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{mod2}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{j1}{Item index (or vector of indices in the case of TIF) for the first model.}

\item{j2}{Item index (or vector of indices in the case of TIF) for the second model.}

\item{method}{Which type of discrepancy to compute? RIMSD, ISD, or IAD.}

\item{theta}{Grid over which to compute the information function or TIF.}

\item{w}{Vector of weights; usually these should sum to 1.}
}
\description{
Difference in information functions or TIF
}
\details{
This is a convenience function which allows computation of RIMSD, ISD, or IAD for item or test information (TIF) directly from two fitted models.
}
