% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{infoMP}
\alias{infoMP}
\title{Computes item information for an item from a fitted MP model.}
\usage{
infoMP(x, j, theta = seq(-5, 5, 0.1))
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}}.}

\item{j}{Item index.}

\item{theta}{Grid for theta, the latent trait.}
}
\description{
Computes item information for an item from a fitted MP model.
}
\details{
Extracts the given item and then computes and returns item information along the grid specified by theta.
}
