% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{infoMx}
\alias{infoMx}
\title{Computes item information}
\usage{
infoMx(x, j, theta = seq(-5, 5, 0.1))
}
\arguments{
\item{x}{- fitted mxModel}

\item{j}{- item index}

\item{theta}{- grid for theta}
}
\description{
Computes item information
}
