% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{newkmat}
\alias{newkmat}
\title{Function to create matrix that represents "k" for each item}
\usage{
newkmat(kstart, kmax, ni)
}
\arguments{
\item{kstart}{Starting value for k for all items}

\item{kmax}{Maximum k}

\item{ni}{Number of items}
}
\description{
Function to create matrix that represents "k" for each item
}
