% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPsearchfunctions.R
\name{newTemp}
\alias{newTemp}
\title{Computes algorithm "temperature" based on initial temperature, max number of iterations, current iteration, and last temperature}
\usage{
newTemp(
  inittemp,
  itermax,
  i,
  lasttemp = 1e-05,
  type = c("straight", "straight2", "logarithmic")
)
}
\arguments{
\item{inittemp}{Initial temperature.}

\item{itermax}{Maximum number of iterations.}

\item{i}{Current iteration}

\item{lasttemp}{Last temperature (i.e., when i == itermax )}

\item{type}{Type of temperature schedule}
}
\description{
Computes algorithm "temperature" based on initial temperature, max number of iterations, current iteration, and last temperature
}
