% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{plotInfoMP}
\alias{plotInfoMP}
\title{Plots item information for an item from a fitted MP model.}
\usage{
plotInfoMP(x, j, theta = seq(-5, 5, 0.1), ...)
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}}.}

\item{j}{Item index.}

\item{theta}{Grid for theta, thet latent trait.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\description{
Plots item information for an item from a fitted MP model.
}
