% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{plotInfoMx}
\alias{plotInfoMx}
\title{plots of info}
\usage{
plotInfoMx(x, j, th = seq(-5, 5, 0.1), ...)
}
\arguments{
\item{x}{blah}

\item{j}{blah}

\item{th}{blah}

\item{...}{blah}
}
\description{
plots of info
}
