% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{plotMx}
\alias{plotMx}
\title{Plots traceline}
\usage{
plotMx(x, j, itemtype = "", th = seq(-5, 5, 0.1), col = "black", ...)
}
\arguments{
\item{x}{- is an fitted mxModel}

\item{j}{- item index}

\item{itemtype}{- is the type of item; must be specified here}

\item{th}{grid}

\item{col}{color}

\item{...}{blah}
}
\description{
Plots traceline
}
