% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{plotTraceMP}
\alias{plotTraceMP}
\title{Plots traceline for an item from a fitted MP model.}
\usage{
plotTraceMP(x, j, theta = seq(-5, 5, 0.1), col = "black", ...)
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}}.}

\item{j}{Item index.}

\item{theta}{Grid for theta, thet latent trait.}

\item{col}{Single color or vector of colors to use for the categories.}

\item{...}{Additional options passed to \code{plot}.}
}
\description{
Plots traceline for an item from a fitted MP model.
}
\details{
Uses base R graphics for now.
}
