% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{RIMSD}
\alias{RIMSD}
\title{Root Integrated Mean Square Difference}
\usage{
RIMSD(T1, T2, w)
}
\arguments{
\item{T1}{Vector of values corresponding to traceline or information function 1.}

\item{T2}{Vector of values corresponding to traceline or information function 2.}

\item{w}{Vector of weights; usually these should sum to 1.}
}
\description{
Root Integrated Mean Square Difference
}
\details{
Computes root integrated mean square difference (RIMSD) between \code{T1} and \code{T2}, with weights for integration
from \code{w}. The difference is squared, a weighted sum is computed based on \code{w} and then the square root is taken.
}
