% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPsearchfunctions.R
\name{sim.anneal}
\alias{sim.anneal}
\title{Main function that performs simulated annealing (SA)}
\usage{
sim.anneal(
  dat,
  k.mat,
  itermax = 1000,
  inittemp = 10,
  type = "aic",
  random = FALSE,
  priors = TRUE,
  startimat = NULL,
  step = 1,
  items = 1,
  pvar = 1000,
  taumean = -10,
  itemtype = NULL,
  qpoints = 101,
  qwidth = 5,
  temptype = "straight",
  ...
)
}
\arguments{
\item{dat}{The data, typically in a format prepared by \code{\link[OpenMx]{mxFactor}}.}

\item{k.mat}{Starting matrix that represents k for each item (e.g., as generated by \code{\link{newkmat}}).}

\item{itermax}{Number of iterations for SA algorithm.}

\item{inittemp}{Staring temperature for SA algorithm.}

\item{type}{Which information criterion to use as the objective function to minimize. Anything supported by \code{\link{getIC}} is currently possible.}

\item{random}{Logical value indicating whether to just automatically accept neighboring models (meaning not actually do SA).}

\item{priors}{Logical value indicating whether to use prior distributions on alpha and tau parameters.}

\item{startimat}{Sarting item parameter matrix, if any. Used only for the first iteration.}

\item{step}{Allowed change in k for candidate models (see documentation for compute.neighbor function)}

\item{items}{Number of items for perturbing k (see documentation for compute.neighbor function)}

\item{pvar}{- prior variance}

\item{taumean}{- mean of prior distribution for tau parameters; currently all parameters will have the same prior mean}

\item{itemtype}{- character vector that has length same as number of items; indicates what type of model to use for each item}

\item{qpoints}{- number of quadrature points}

\item{qwidth}{- width for quadrature grid}

\item{temptype}{- method for temperature schedule decreases (see newtemp function)}

\item{...}{Additional arguments to be passed to \code{\link{fitMP}}.}
}
\description{
Main function that performs simulated annealing (SA)
}
