% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{TCC}
\alias{TCC}
\title{Compute test characteristic curve (TCC) for items from a fitted MP model.}
\usage{
TCC(x, items, theta = seq(-5, 5, 0.1), scale = FALSE, scaletest = FALSE)
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}}.}

\item{items}{Vector of indices corresponding to items to include in TCC computations.}

\item{theta}{Grid for theta, the latent trait.}

\item{scale}{Logical value. If TRUE, divide each items' expected score curve by the total number of categories so that the expected sore curve is bound between 0 and 1.}

\item{scaletest}{Logical value. If TRUE, divide the TCC by the total number of items.}
}
\description{
Compute test characteristic curve (TCC) for items from a fitted MP model.
}
