% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{TCCMx}
\alias{TCCMx}
\title{Compute TCC}
\usage{
TCCMx(x, items, theta = seq(-5, 5, 0.1), scale = FALSE, scaletest = FALSE)
}
\arguments{
\item{x}{blah}

\item{items}{blah}

\item{theta}{blah}

\item{scale}{blah}

\item{scaletest}{blah}
}
\description{
Compute TCC
}
