% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{TIF}
\alias{TIF}
\title{Test information function (TIF) for items from a fitted MP model.}
\usage{
TIF(x, items, theta = seq(-5, 5, 0.1))
}
\arguments{
\item{x}{Fitted mxModel (e.g., from \code{\link{fitMP}}.}

\item{items}{Vector of indices corresponding to items to include in TCC computations.}

\item{theta}{Grid for theta, the latent trait.}
}
\description{
Test information function (TIF) for items from a fitted MP model.
}
