% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{TIFMx}
\alias{TIFMx}
\title{test information function}
\usage{
TIFMx(x, items, theta = seq(-5, 5, 0.1))
}
\arguments{
\item{x}{blah}

\item{items}{blah}

\item{theta}{blah}
}
\description{
test information function
}
