% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{tracedif}
\alias{tracedif}
\title{Difference in tracelines or TCC}
\usage{
tracedif(
  mod1,
  mod2,
  j1,
  j2,
  method = c("RIMSD", "ISD", "IAD"),
  scale = TRUE,
  scaletest = TRUE,
  theta = seq(-5, 5, 0.1),
  w = dnorm(theta)/sum(dnorm(theta))
)
}
\arguments{
\item{mod1}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{mod2}{Fitted \code{mxModel}, e.g., from \code{\link{fitMP}}.}

\item{j1}{Item index (or vector of indices in the case of TCC) for the first model.}

\item{j2}{Item index (or vector of indices in the case of TCC) for the second model.}

\item{method}{Which type of discrepancy to compute? RIMSD, ISD, or IAD.}

\item{scale}{logical value to determine whether to scale each expected score value to 0-1. See \code{\link{TCC}}.}

\item{scaletest}{logical value to determine whether to scale the overall test by dividing by the number of items. See \code{\link{TCC}}.}

\item{theta}{Grid over which to compute the traceline or TCC.}

\item{w}{Vector of weights; usually these should sum to 1.}
}
\description{
Difference in tracelines or TCC
}
\details{
This is a convenience function which allows computation of RIMSD, ISD, or IAD directly from two fitted models
  and will do so on either single items from each or for TCCs if a vector of items from each is selected. Note that RIMSD
  is usually scaled so that the resulting value is on a metric between 0 and 1. Variants of ISD and IAD (wABC an sDTF) in previous
  papers were dependent on the number of score categories. However, Chalmers et al. mentions a variant that can serve as an effect
  size if further scaled.
}
