% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPdifferencefunctions.R
\name{tracedifitem}
\alias{tracedifitem}
\title{Difference in tracelines for just a pair of items}
\usage{
tracedifitem(
  item1,
  item2,
  par1,
  par2,
  method = c("RIMSD", "ISD", "IAD"),
  scale = TRUE,
  theta = seq(-5, 5, 0.1),
  w = dnorm(theta)/sum(dnorm(theta))
)
}
\arguments{
\item{item1}{Item specification. e.g., from \code{rpf}.}

\item{item2}{Item specification. e.g., from \code{rpf}.}

\item{par1}{Item parameters for item 1.}

\item{par2}{Item parameters for item 2.}

\item{method}{Which type of discrepancy to compute? RIMSD, ISD, or IAD.}

\item{scale}{logical value to determine whether to scale each expected score value to 0-1.}

\item{theta}{Grid over which to compute the traceline.}

\item{w}{Vector of weights; usually these should sum to 1.}
}
\description{
Difference in tracelines for just a pair of items
}
\details{
This is a convenience function which allows computation of RIMSD, ISD, or IAD from two item specifications and their parameters.
}
