% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPfunctions.R
\name{traceProbMx}
\alias{traceProbMx}
\title{Extracts T(theta), or response functions, for an item from a fitted mxModel
and for all categories - one category per column}
\usage{
traceProbMx(x, j, itemtype = "", theta = seq(-5, 5, 0.1))
}
\arguments{
\item{x}{- fitted mxModel}

\item{j}{- item index}

\item{itemtype}{- type of item (lmp, grmp, gpcmp)}

\item{theta}{- grid for theta}
}
\description{
Extracts T(theta), or response functions, for an item from a fitted mxModel
and for all categories - one category per column
}
