% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPsearchfunctions.R
\name{getIC}
\alias{getIC}
\title{To extract an information criterion from a fitted model}
\usage{
getIC(fitmodel, type = "aic", N, usefitfunc = FALSE, infotype = "oakes1999")
}
\arguments{
\item{fitmodel}{blah}

\item{type}{blah}

\item{N}{blah}

\item{usefitfunc}{blah}

\item{infotype}{blah}
}
\description{
To extract an information criterion from a fitted model
}
