% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPsearchfunctions.R
\name{sim.anneal}
\alias{sim.anneal}
\title{Main function that performs simulated annealing}
\usage{
sim.anneal(
  dat,
  k.mat,
  itermax = 1000,
  inittemp = 10,
  type = "aic",
  random = FALSE,
  priors = TRUE,
  startimat = NULL,
  step = 1,
  items = 1,
  pvar = 1000,
  taumean = -10,
  itemtype = NULL,
  qpoints = 101,
  qwidth = 5,
  temptype = "straight",
  ...
)
}
\arguments{
\item{dat}{- data matrix of item responses}

\item{k.mat}{- matrix that represents k for each item}

\item{itermax}{- maximum number of iterations}

\item{inittemp}{- staring temperature for SA}

\item{type}{- blah}

\item{random}{- logical value indicating whether to just automatically accept neighboring models (meaning not actually do SA)}

\item{priors}{- logical value indicating whether to use prior distributions on alpha and tau parameters}

\item{startimat}{- starting item parameter matrix, if any. For example, say we start with a model with k=1 for all items}

\item{step}{- allowed change in k for candidate models (see documentation for compute.neighbor function)}

\item{items}{- number of items for perturbing k (see documentation for compute.neighbor function)}

\item{pvar}{- prior variance}

\item{taumean}{- mean of prior distribution for tau parameters; currently all parameters will have the same prior mean}

\item{itemtype}{- character vector that has length same as number of items; indicates what type of model to use for each item}

\item{qpoints}{- number of quadrature points}

\item{qwidth}{- width for quadrature grid}

\item{temptype}{- method for temperature schedule decreases (see newtemp function)}

\item{...}{blah}
}
\description{
Main function that performs simulated annealing
}
