% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPsearchfunctions.R
\name{simAnneal}
\alias{simAnneal}
\title{Main function that performs simulated annealing (SA)}
\usage{
simAnneal(
  dat,
  k.mat,
  itermax = 1000,
  inittemp = 10,
  type = "aic",
  random = FALSE,
  startimat = NULL,
  step = 1,
  items = 1,
  temptype = "straight",
  ...
)
}
\arguments{
\item{dat}{The data, typically in a format prepared by \code{\link[OpenMx]{mxFactor}}.}

\item{k.mat}{Starting matrix that represents k for each item (e.g., as generated by \code{\link{newkmat}}).}

\item{itermax}{Number of iterations for SA algorithm.}

\item{inittemp}{Staring temperature for SA algorithm.}

\item{type}{Which information criterion to use as the objective function to minimize. Anything supported by \code{\link{getIC}} is currently possible.}

\item{random}{Logical value indicating whether to just automatically accept neighboring models (meaning not actually do SA).}

\item{startimat}{Starting item parameter matrix, if any. Used only for the first iteration.}

\item{step}{Maximum allowed change in k for candidate models. e.g., step 1 -> k=0 to k=1 allowed, but not k=0 to k=2. Smaller steps less likely to have estimation problems.}

\item{items}{How many items' k should be perturbed for computing a neighboring model?}

\item{temptype}{Method for temperature schedule decreases (see \code{\link{newTemp}} function).}

\item{...}{Additional arguments to be passed to \code{\link{fitMP}}.}
}
\value{
A list with the following elements
}
\description{
Main function that performs simulated annealing (SA)
}
\details{
This function performs simulated annealing to select polynomial order.
}
